refactor(resolver): decompose plugin resolution architecture
