+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/LocalPluginJar.kt
@@ -1,0 +1,5 @@
+package me.liam.microsmith.cli.plugins
+
+import java.nio.file.Path
+
+internal data class LocalPluginJar(val artifactPath: Path, val lockKey: String)
