+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/LocalPluginResolution.kt
@@ -1,0 +1,5 @@
+package me.liam.microsmith.cli.plugins
+
+import java.nio.file.Path
+
+internal data class LocalPluginResolution(val classpath: List<Path>, val lockEntries: List<LockEntry>)
