+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/LockEntry.kt
@@ -1,0 +1,3 @@
+package me.liam.microsmith.cli.plugins
+
+internal data class LockEntry(val kind: String, val key: String, val checksum: String)
