+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/LockKey.kt
@@ -1,0 +1,3 @@
+package me.liam.microsmith.cli.plugins
+
+internal data class LockKey(val kind: String, val key: String)
