+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/MavenDependencyGraphRequest.kt
@@ -1,0 +1,11 @@
+package me.liam.microsmith.cli.plugins
+
+import org.eclipse.aether.repository.RemoteRepository
+import java.nio.file.Path
+
+internal data class MavenDependencyGraphRequest(
+    val coordinate: Coordinate,
+    val repositories: List<RemoteRepository>,
+    val localRepositoryRoot: Path,
+    val offline: Boolean,
+)
