+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/ParsedLockfile.kt
@@ -1,0 +1,3 @@
+package me.liam.microsmith.cli.plugins
+
+internal data class ParsedLockfile(val version: Int, val entries: List<LockEntry>)
