--- cli/src/main/kotlin/me/liam/microsmith/cli/plugins/PluginRepository.kt
@@ -1,53 +1,0 @@
-package me.liam.microsmith.cli.plugins
-
-import me.liam.microsmith.cli.command.RunCommand
-import java.nio.file.Path
-
-internal fun resolveRepositories(command: RunCommand, settings: PluginResolverSettings): List<String> =
-    resolveRepositories(command, settings, settings.repositoryPolicy ?: defaultRepositoryAllowlistPolicy())
-
-internal fun resolveRepositories(
-    command: RunCommand,
-    settings: PluginResolverSettings,
-    repositoryPolicy: RepositoryAllowlistPolicy,
-): List<String> {
-    val override = command.repositoryOverride?.trim()?.takeIf { it.isNotEmpty() }
-    val repositories =
-        (listOfNotNull(override) + settings.defaultRepositories)
-            .map(::normalizeRepositoryUri)
-            .distinct()
-    repositories.forEach(repositoryPolicy::validate)
-    return repositories
-}
-
-internal fun resolveRepositoryEndpoints(
-    command: RunCommand,
-    settings: PluginResolverSettings,
-    repositoryPolicy: RepositoryAllowlistPolicy,
-    credentialsResolver: RepositoryCredentialsResolver,
-): List<RepositoryEndpoint> = resolveRepositories(command, settings, repositoryPolicy)
-    .map { repositoryUri ->
-        RepositoryEndpoint(
-            uri = repositoryUri,
-            credentials = credentialsResolver.resolve(repositoryUri),
-        )
-    }
-
-internal fun pluginArtifactCacheRoot(cacheDirectory: Path): Path = cacheDirectory
-    .resolve("artifacts")
-    .toAbsolutePath()
-    .normalize()
-
-internal fun cachePathFor(cacheRoot: Path, coordinate: Coordinate): Path {
-    val artifactPath =
-        cacheRoot
-            .resolve(coordinate.group.replace('.', '/'))
-            .resolve(coordinate.artifact)
-            .resolve(coordinate.version)
-            .resolve("${coordinate.artifact}-${coordinate.version}.jar")
-            .normalize()
-    require(artifactPath.startsWith(cacheRoot)) {
-        "Plugin coordinate '${coordinate.value}' resolves outside plugin cache directory."
-    }
-    return artifactPath
-}
