+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/PluginResolutionConstants.kt
@@ -1,0 +1,7 @@
+package me.liam.microsmith.cli.plugins
+
+internal const val LOCKFILE_VERSION = 2
+internal const val MAVEN_CENTRAL_REPOSITORY = "https://repo1.maven.org/maven2"
+internal const val REMOTE_KIND = "remote"
+internal const val REMOTE_ARTIFACT_KIND = "remote-artifact"
+internal const val LOCAL_KIND = "local"
