+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/PluginResolutionContext.kt
@@ -1,0 +1,11 @@
+package me.liam.microsmith.cli.plugins
+
+import java.nio.file.Path
+
+internal data class PluginResolutionContext(
+    val lockfilePath: Path,
+    val lockfile: ParsedLockfile?,
+    val checksumAllowlist: PluginChecksumAllowlist?,
+    val cacheDirectory: Path,
+    val repositories: List<RepositoryEndpoint>,
+)
