+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/PluginResolutionDiagnosticException.kt
@@ -1,0 +1,7 @@
+package me.liam.microsmith.cli.plugins
+
+internal class PluginResolutionDiagnosticException(
+    val category: PluginResolverErrorCategory,
+    message: String,
+    cause: Throwable? = null,
+) : RuntimeException(message, cause)
