+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/PluginResolutionDiagnostics.kt
@@ -1,0 +1,15 @@
+package me.liam.microsmith.cli.plugins
+
+internal class PluginResolutionDiagnostics {
+    fun format(error: Throwable, sensitiveValues: Set<String>): String = when (error) {
+        is PluginResolutionDiagnosticException ->
+            "[${error.category.code}] " +
+                (error.message ?: "plugin resolution failed")
+                    .redactSensitiveValues(sensitiveValues)
+
+        else -> {
+            val unexpectedMessage = error.message ?: error::class.simpleName ?: "unknown plugin resolution error"
+            "[unexpected] ${unexpectedMessage.redactSensitiveValues(sensitiveValues)}"
+        }
+    }
+}
