+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/PluginResolverErrorCategory.kt
@@ -1,0 +1,10 @@
+package me.liam.microsmith.cli.plugins
+
+internal enum class PluginResolverErrorCategory(val code: String) {
+    OFFLINE_CACHE_MISS("offline-cache-miss"),
+    AUTHENTICATION("authentication"),
+    DEPENDENCY_RESOLUTION("dependency-resolution"),
+    LOCKFILE("lockfile"),
+    REPOSITORY_POLICY("repository-policy"),
+    ROOT_ARTIFACT_MISSING("root-artifact-missing"),
+}
