+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/RemotePluginResolution.kt
@@ -1,0 +1,9 @@
+package me.liam.microsmith.cli.plugins
+
+import java.nio.file.Path
+
+internal data class RemotePluginResolution(
+    val classpath: List<Path>,
+    val rootLockEntries: List<LockEntry>,
+    val remoteArtifactChecksums: Map<String, String>,
+)
