+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/RepositoryCredentials.kt
@@ -1,0 +1,3 @@
+package me.liam.microsmith.cli.plugins
+
+internal data class RepositoryCredentials(val username: String, val password: String)
