+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/ResolvedRemoteArtifact.kt
@@ -1,0 +1,5 @@
+package me.liam.microsmith.cli.plugins
+
+import java.nio.file.Path
+
+internal data class ResolvedRemoteArtifact(val lockKey: String, val artifactPath: Path)
