+++ cli/src/main/kotlin/me/liam/microsmith/cli/plugins/ResolvedRemotePlugin.kt
@@ -1,0 +1,9 @@
+package me.liam.microsmith.cli.plugins
+
+import java.nio.file.Path
+
+internal data class ResolvedRemotePlugin(
+    val rootArtifactPath: Path,
+    val classpath: List<Path>,
+    val artifacts: List<ResolvedRemoteArtifact>,
+)
